+++ AffindaAPI/AffindaAPI/Models/FloatAnnotation.Serialization.cs
@@ -11,15 +11,22 @@ using Azure.Core;
 
 namespace Affinda.API.Models
 {
-    public partial class RowAnnotation : IUtf8JsonSerializable
+    public partial class FloatAnnotation : IUtf8JsonSerializable
     {
         void IUtf8JsonSerializable.Write(Utf8JsonWriter writer)
         {
             writer.WriteStartObject();
             if (Optional.IsDefined(Parsed))
             {
-                writer.WritePropertyName("parsed");
-                writer.WriteObjectValue(Parsed);
+                if (Parsed != null)
+                {
+                    writer.WritePropertyName("parsed");
+                    writer.WriteNumberValue(Parsed.Value);
+                }
+                else
+                {
+                    writer.WriteNull("parsed");
+                }
             }
             writer.WritePropertyName("id");
             writer.WriteNumberValue(Id);
@@ -114,9 +121,9 @@ namespace Affinda.API.Models
             writer.WriteEndObject();
         }
 
-        internal static RowAnnotation DeserializeRowAnnotation(JsonElement element)
+        internal static FloatAnnotation DeserializeFloatAnnotation(JsonElement element)
         {
-            Optional<RowAnnotationParsed> parsed = default;
+            Optional<float?> parsed = default;
             int id = default;
             Rectangle rectangle = default;
             IList<Rectangle> rectangles = default;
@@ -139,10 +146,10 @@ namespace Affinda.API.Models
                 {
                     if (property.Value.ValueKind == JsonValueKind.Null)
                     {
-                        property.ThrowNonNullablePropertyIsNull();
+                        parsed = null;
                         continue;
                     }
-                    parsed = RowAnnotationParsed.DeserializeRowAnnotationParsed(property.Value);
+                    parsed = property.Value.GetSingle();
                     continue;
                 }
                 if (property.NameEquals("id"))
@@ -258,7 +265,7 @@ namespace Affinda.API.Models
                 additionalPropertiesDictionary.Add(property.Name, property.Value.GetObject());
             }
             additionalProperties = additionalPropertiesDictionary;
-            return new RowAnnotation(id, rectangle, rectangles, document.Value, pageIndex, raw, confidence, classificationConfidence, textExtractionConfidence, isVerified, isClientVerified, isAutoVerified, dataPoint, contentType, additionalProperties, parsed.Value);
+            return new FloatAnnotation(id, rectangle, rectangles, document.Value, pageIndex, raw, confidence, classificationConfidence, textExtractionConfidence, isVerified, isClientVerified, isAutoVerified, dataPoint, contentType, additionalProperties, Optional.ToNullable(parsed));
         }
     }
 }